refactor: 컨트롤러에서 도메인을 받지 않도록 재구현
이제 도메인은 서비스에서만 받도록 하였고
컨트롤러는 dto 정보만 받도록 하였다.
